requirements.txt
requests
pyyaml
prettytable
python-dotenv
sentence-transformers
scikit-learn
jinja2